fix(workspaces): use regular integer, its fine for now...

+++ packages/server/modules/workspacesCore/migrations/20250822161233_workspace_feature_flags.ts
@@ -2,7 +2,7 @@ import { type Knex } from 'knex'
 
 export async function up(knex: Knex): Promise<void> {
   await knex.schema.alterTable('workspace_plans', (table) => {
-    table.bigInteger('featureFlags').notNullable().defaultTo(0)
+    table.integer('featureFlags').notNullable().defaultTo(0)
   })
 }
